rename method and file names for gcpApis
This commit aims to improve code clarity by renaming
methods and files related to GCP APIs
to more descriptive names.
The changes include renaming multipart upload methods
and object tagging methods to better reflect their
functionality.
Enhancing code readability and maintainability.
Issue: ARSN-524
