+++ lib/storage/data/external/GCP/GcpApis/abortMultipartUpload.js
@@ -13,7 +13,7 @@ const { logHelper } = require('../../utils');
  * @param {function} callback - callback function to call
  * @return {undefined}
  */
-function abortMPU(params, callback) {
+function abortMultipartUpload(params, callback) {
     if (!params || !params.Key || !params.UploadId ||
         !params.Bucket || !params.MPU) {
         const error = errorInstances.InvalidRequest
@@ -30,4 +30,4 @@ function abortMPU(params, callback) {
     return mpuHelper.removeParts(delParams, callback);
 }
 
-module.exports = abortMPU;
+module.exports = abortMultipartUpload;
