+++ lib/storage/data/external/GCP/GcpApis/completeMultipartUpload.js
@@ -4,7 +4,6 @@ const errorInstances = require('../../../../../errors').errorInstances;
 const MpuHelper = require('./mpuHelper');
 const { createMpuKey, logger } = require('../GcpUtils');
 const { logHelper } = require('../../utils');
-
 /**
  * completeMPU - merges a list of parts into a single object
  * @param {object} params - completeMPU params
@@ -17,7 +16,7 @@ const { logHelper } = require('../../utils');
  * @param {function} callback - callback function to call with MPU result
  * @return {undefined}
  */
-function completeMPU(params, callback) {
+function completeMultipartUpload(params, callback) {
     if (!params || !params.MultipartUpload ||
         !params.MultipartUpload.Parts || !params.UploadId ||
         !params.Bucket || !params.Key) {
@@ -77,4 +76,4 @@ function completeMPU(params, callback) {
     });
 }
 
-module.exports = completeMPU;
+module.exports = completeMultipartUpload;
