+++ lib/storage/data/external/GCP/GcpApis/createMultipartUpload.js
@@ -18,7 +18,7 @@ const { logHelper } = require('../../utils');
  * upload-id for MPU operations
  * @return {undefined}
  */
-function createMPU(params, callback) {
+function createMultipartUpload(params, callback) {
     // As google cloud does not have a create MPU function,
     // create an empty 'init' object that will temporarily store the
     // object metadata and return an upload ID to mimic an AWS MPU
@@ -32,7 +32,6 @@ function createMPU(params, callback) {
     const mpuParams = {
         Bucket: params.Bucket,
         Key: createMpuKey(params.Key, uploadId, 'init'),
-        Body: Buffer.alloc(0),
         Metadata: params.Metadata,
         ContentType: params.ContentType,
         CacheControl: params.CacheControl,
@@ -49,4 +48,4 @@ function createMPU(params, callback) {
     });
 }
 
-module.exports = createMPU;
+module.exports = createMultipartUpload;
