+++ lib/storage/data/external/GCP/GcpApis/deleteObjectTagging.js
