+++ lib/storage/data/external/GCP/GcpApis/getObjectTagging.js
