+++ lib/storage/data/external/GCP/GcpApis/index.js
@@ -1,14 +1,14 @@
 module.exports = {
     // mpu functions
-    abortMultipartUpload: require('./abortMPU'),
-    completeMultipartUpload: require('./completeMPU'),
-    createMultipartUpload: require('./createMPU'),
+    abortMultipartUpload: require('./abortMultipartUpload'),
+    completeMultipartUpload: require('./completeMultipartUpload'),
+    createMultipartUpload: require('./createMultipartUpload'),
     listParts: require('./listParts'),
     uploadPart: require('./uploadPart'),
     uploadPartCopy: require('./uploadPartCopy'),
     // object tagging
     putObject: require('./putObject'),
-    putObjectTagging: require('./putTagging'),
-    getObjectTagging: require('./getTagging'),
-    deleteObjectTagging: require('./deleteTagging'),
+    putObjectTagging: require('./putObjectTagging'),
+    getObjectTagging: require('./getObjectTagging'),
+    deleteObjectTagging: require('./deleteObjectTagging'),
 };
